requirements.txt
pytest
selenium == 4.11.2
selene >= 2.0.0b17
allure-pytest
python-dotenv